+++ test/AccessManager.t.sol
@@ -74,7 +74,7 @@ contract AccessManagerTest is Test {
         uint amount = 10000e8;
         res.underlying.mint(deployer, amount);
         res.underlying.approve(address(res.market), amount);
-        res.market.provideLiquidity(amount);
+        res.market.provideLiquidity(uint128(amount));
 
         vm.stopPrank();
 
@@ -82,7 +82,7 @@ contract AccessManagerTest is Test {
 
         vm.startPrank(sender);
         res.underlying.approve(address(res.market), amount);
-        res.market.provideLiquidity(amount);
+        res.market.provideLiquidity(uint128(amount));
         vm.stopPrank();
     }
